+++ pika/adapters/tornado_connection.py
@@ -1,8 +1,11 @@
 """Run pika on the Tornado IOLoop"""
 from tornado import ioloop
+import logging
 import time
 
 from pika.adapters import base_connection
+
+LOGGER = logging.getLogger(__name__)
 
 
 class TornadoConnection(base_connection.BaseConnection):
@@ -16,20 +19,45 @@ class TornadoConnection(base_connection.BaseConnection):
 
     def __init__(self, parameters=None,
                  on_open_callback=None,
+                 on_open_error_callback=None,
+                 on_close_callback=None,
                  stop_ioloop_on_close=False,
                  custom_ioloop=None):
-        self._ioloop = custom_ioloop or ioloop.IOLoop.instance()
-        super(TornadoConnection, self).__init__(parameters, on_open_callback,
+        """Create a new instance of the TornadoConnection class, connecting
+        to RabbitMQ automatically
+
+        :param pika.connection.Parameters parameters: Connection parameters
+        :param on_open_callback: The method to call when the connection is open
+        :type on_open_callback: method
+        :param on_open_error_callback: Method to call if the connection cant
+                                       be opened
+        :type on_open_error_callback: method
+        :param bool stop_ioloop_on_close: Call ioloop.stop() if disconnected
+        :param custom_ioloop: Override using the global IOLoop in Tornado
+
+        """
+        self.sleep_counter = 0
+        self.ioloop = custom_ioloop or ioloop.IOLoop.instance()
+        super(TornadoConnection, self).__init__(parameters,
+                                                on_open_callback,
+                                                on_open_error_callback,
+                                                on_close_callback,
+                                                self.ioloop,
                                                 stop_ioloop_on_close)
 
     def _adapter_connect(self):
-        """Connect to the RabbitMQ broker"""
-        super(TornadoConnection, self)._adapter_connect()
-        self.ioloop = self._ioloop
-        self.ioloop.add_handler(self.socket.fileno(),
-                                self._handle_events,
-                                self.event_state)
-        self._on_connected()
+        """Connect to the remote socket, adding the socket to the IOLoop if
+        connected
+
+        :rtype: bool
+
+        """
+        if super(TornadoConnection, self)._adapter_connect():
+            self.ioloop.add_handler(self.socket.fileno(),
+                                    self._handle_events,
+                                    self.event_state)
+            return True
+        return False
 
     def _adapter_disconnect(self):
         """Disconnect from the RabbitMQ broker"""