+++ pika/data.py
@@ -21,6 +21,8 @@ def encode_table(pieces, table):
     pieces.append(None)  # placeholder
     tablesize = 0
     for (key, value) in table.iteritems():
+        if isinstance(key, unicode):
+            key = key.encode('utf-8')
         pieces.append(struct.pack('B', len(key)))
         pieces.append(key)
         tablesize = tablesize + 1 + len(key)
@@ -40,7 +42,8 @@ def encode_value(pieces, value):
 
     """
     if isinstance(value, basestring):
-        value = value.encode('utf8')
+        if isinstance(value, unicode):
+            value = value.encode('utf-8')
         pieces.append(struct.pack('>cI', 'S', len(value)))
         pieces.append(value)
         return 5 + len(value)
@@ -120,15 +123,72 @@ def decode_value(encoded, offset):
     """
     kind = encoded[offset]
     offset += 1
-    if kind == 'S':
-        length = struct.unpack_from('>I', encoded, offset)[0]
+
+    # Bool
+    if kind == 't':
+        value = struct.unpack_from('>B', encoded, offset)[0]
+        value = bool(value)
+        offset += 1
+
+    # Short-Short Int
+    elif kind == 'b':
+        value = struct.unpack_from('>B', encoded, offset)[0]
+        offset += 1
+
+    # Short-Short Unsigned Int
+    elif kind == 'B':
+        value = struct.unpack_from('>b', encoded, offset)[0]
+        offset += 1
+
+    # Short Int
+    elif kind == 'U':
+        value = struct.unpack_from('>h', encoded, offset)[0]
+        offset += 2
+
+    # Short Unsigned Int
+    elif kind == 'u':
+        value = struct.unpack_from('>H', encoded, offset)[0]
+        offset += 2
+
+    # Long Int
+    elif kind == 'I':
+        value = struct.unpack_from('>i', encoded, offset)[0]
         offset += 4
-        value = encoded[offset: offset + length].decode('utf8')
-        try:
-            value = str(value)
-        except UnicodeEncodeError:
-            pass
-        offset += length
+
+    # Long Unsigned Int
+    elif kind == 'i':
+        value = struct.unpack_from('>I', encoded, offset)[0]
+        offset += 4
+
+    # Long-Long Int
+    elif kind == 'L':
+        value = long(struct.unpack_from('>q', encoded, offset)[0])
+        offset += 8
+
+    # Long-Long Unsigned Int
+    elif kind == 'l':
+        value = long(struct.unpack_from('>Q', encoded, offset)[0])
+        offset += 8
+
+    # Float
+    elif kind == 'f':
+        value = long(struct.unpack_from('>f', encoded, offset)[0])
+        offset += 4
+
+    # Double
+    elif kind == 'd':
+        value = long(struct.unpack_from('>d', encoded, offset)[0])
+        offset += 8
+
+    # Decimal
+    elif kind == 'D':
+        decimals = struct.unpack_from('B', encoded, offset)[0]
+        offset += 1
+        raw = struct.unpack_from('>i', encoded, offset)[0]
+        offset += 4
+        value = decimal.Decimal(raw) * (decimal.Decimal(10) ** -decimals)
+
+    # Short String
     elif kind == 's':
         length = struct.unpack_from('B', encoded, offset)[0]
         offset += 1
@@ -138,28 +198,19 @@ def decode_value(encoded, offset):
         except UnicodeEncodeError:
             pass
         offset += length
-    elif kind == 't':
-        value = struct.unpack_from('>B', encoded, offset)[0]
-        value = bool(value)
-        offset += 1
-    elif kind == 'I':
-        value = struct.unpack_from('>i', encoded, offset)[0]
+
+    # Long String
+    elif kind == 'S':
+        length = struct.unpack_from('>I', encoded, offset)[0]
         offset += 4
-    elif kind == 'l':
-        value = long(struct.unpack_from('>q', encoded, offset)[0])
-        offset += 8
-    elif kind == 'D':
-        decimals = struct.unpack_from('B', encoded, offset)[0]
-        offset += 1
-        raw = struct.unpack_from('>i', encoded, offset)[0]
-        offset += 4
-        value = decimal.Decimal(raw) * (decimal.Decimal(10) ** -decimals)
-    elif kind == 'T':
-        value = datetime.utcfromtimestamp(struct.unpack_from('>Q', encoded,
-                                                             offset)[0])
-        offset += 8
-    elif kind == 'F':
-        (value, offset) = decode_table(encoded, offset)
+        value = encoded[offset: offset + length].decode('utf8')
+        try:
+            value = str(value)
+        except UnicodeEncodeError:
+            pass
+        offset += length
+
+    # Field Array
     elif kind == 'A':
         length = struct.unpack_from('>I', encoded, offset)[0]
         offset += 4
@@ -168,8 +219,21 @@ def decode_value(encoded, offset):
         while offset < offset_end:
             v, offset = decode_value(encoded, offset)
             value.append(v)
+
+    # Timestamp
+    elif kind == 'T':
+        value = datetime.utcfromtimestamp(struct.unpack_from('>Q', encoded,
+                                                             offset)[0])
+        offset += 8
+
+    # Field Table
+    elif kind == 'F':
+        (value, offset) = decode_table(encoded, offset)
+
+    # Null / Void
     elif kind == 'V':
         value = None
     else:
         raise exceptions.InvalidFieldTypeException(kind)
+
     return value, offset