+++ pika/heartbeat.py
@@ -36,6 +36,9 @@ class HeartbeatChecker(object):
         self._heartbeat_frames_sent = 0
         self._idle_byte_intervals = 0
 
+        # The handle for the last timer
+        self._timer = None
+
         # Setup the timer to fire in _interval seconds
         self._setup_timer()
 
@@ -99,13 +102,22 @@ class HeartbeatChecker(object):
         # Update the timer to fire again
         self._start_timer()
 
+    def stop(self):
+        """Stop the heartbeat checker"""
+        if self._timer:
+            LOGGER.debug('Removing timeout for next heartbeat interval')
+            self._connection.remove_timeout(self._timer)
+            self._timer = None
+
     def _close_connection(self):
         """Close the connection with the AMQP Connection-Forced value."""
         LOGGER.info('Connection is idle, %i stale byte intervals',
                     self._idle_byte_intervals)
         duration = self._max_idle_count * self._interval
-        self._connection.close(HeartbeatChecker._CONNECTION_FORCED,
-                               HeartbeatChecker._STALE_CONNECTION % duration)
+        text = HeartbeatChecker._STALE_CONNECTION % duration
+        self._connection.close(HeartbeatChecker._CONNECTION_FORCED, text)
+        self._connection._on_disconnect(HeartbeatChecker._CONNECTION_FORCED,
+                                        text)
 
     @property
     def _has_received_data(self):
@@ -138,7 +150,8 @@ class HeartbeatChecker(object):
         every interval seconds.
 
         """
-        self._connection.add_timeout(self._interval, self.send_and_check)
+        self._timer = self._connection.add_timeout(self._interval,
+                                                   self.send_and_check)
 
     def _start_timer(self):
         """If the connection still has this object set for heartbeats, add a