+++ docs/version_history.rst
@@ -1,7 +1,59 @@
 Version History
 ===============
 
-0.9.9 - Unreleased
+0.9.12 - 2013-03-18
+-------------------
+
+**Bugfixes**
+
+- New timeout id hashing was not unique
+
+0.9.11 - 2013-03-17
+-------------------
+
+**Bugfixes**
+
+- Address inconsistent channel close callback documentation and add the signature
+  change to the TwistedChannel class (#305)
+- Address a missed timeout related internal data structure name change
+  introduced in the SelectConnection 0.9.10 release. Update all connection
+  adapters to use same signature and docstring (#306).
+
+0.9.10 - 2013-03-16
+-------------------
+
+**Bugfixes**
+
+- Fix timeout in twisted adapter (Submitted by cellscape)
+- Fix blocking_connection poll timer resolution to milliseconds (Submitted by cellscape)
+- Fix channel._on_close() without a method frame (Submitted by Richard Boulton)
+- Addressed exception on close (Issue #279 - fix by patcpsc)
+- 'messages' not initialized in BlockingConnection.cancel() (Issue #289 - fix by Mik Kocikowski)
+- Make queue_unbind behave like queue_bind (Issue #277)
+- Address closing behavioral issues for connections and channels (Issue #275)
+- Pass a Method frame to Channel._on_close in Connection._on_disconnect (Submitted by Jan Urbański)
+- Fix channel closed callback signature in the Twisted adapter (Submitted by Jan Urbański)
+- Don't stop the IOLoop on connection close for in the Twisted adapter (Submitted by Jan Urbański)
+- Update the asynchronous examples to fix reconnecting and have it work
+- Warn if the socket was closed such as if RabbitMQ dies without a Close frame
+- Fix URLParameters ssl_options (Issue #296)
+- Add state to BlockingConnection addressing (Issue #301)
+- Encode unicode body content prior to publishing (Issue #282)
+- Fix an issue with unicode keys in BasicProperties headers key (Issue #280)
+- Change how timeout ids are generated (Issue #254)
+- Address post close state issues in Channel (Issue #302)
+
+** Behavior changes **
+
+- Change core connection communication behavior to prefer outbound writes over reads, addressing a recursion issue
+- Update connection on close callbacks, changing callback method signature
+- Update channel on close callbacks, changing callback method signature
+- Give more info in the ChannelClosed exception
+- Change the constructor signature for BlockingConnection, block open/close callbacks
+- Disable the use of add_on_open_callback/add_on_close_callback methods in BlockingConnection
+
+
+0.9.9 - 2013-01-29
 ------------------
 
 **Bugfixes**
@@ -10,7 +62,7 @@ Version History
 - Allow messages with no body (Issue #227)
 - Allow for empty routing keys (Issue #224)
 - Don't raise an exception when trying to send a frame to a closed connection (Issue #229)
-- Only send a Connection.CloseOk if the connection is still open. (Issue #236 - Fix by "noleaf")
+- Only send a Connection.CloseOk if the connection is still open. (Issue #236 - Fix by noleaf)
 - Fix timeout threshold in blocking connection - (Issue #232 - Fix by Adam Flynn)
 - Fix closing connection while a channel is still open (Issue #230 - Fix by Adam Flynn)
 - Fixed misleading warning and exception messages in BaseConnection (Issue #237 - Fix by Tristan Penman)
@@ -21,6 +73,8 @@ Version History
 - Change the string validation to basestring for host connection parameters (Issue #231)
 - Add a poller to the BlockingConnection to address latency issues introduced in Pika 0.9.8 (Issue #242)
 - reply_code and reply_text is not set in ChannelException (Issue #250)
+- Add the missing constraint parameter for Channel._on_return callback processing (Issue #257 - Fix by patcpsc)
+- Channel callbacks not being removed from callback manager when channel is closed or deleted (Issue #261)
 
 0.9.8 - 2012-11-18
 ------------------